+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Command 'deploy list'
 - Command 'team remove-user'
 - Command 'deploy rollback'
+- Command 'app delete'
+- Command 'deploy rollback'
 
 ### Changed
 - Deploys are now performed using the 'deploy create' command